fixed compile bug and changed ulong to uint32_t

+++ backend/include/Helper.h
@@ -37,7 +37,7 @@ std::string generateUniqueHash();
 int randomNumberInRange(int min, int max);
 std::string randomString(const int len);
 std::string randomNDigitNumber(int n);
-std::string generateVerificationEmailText(ulong puid, std::string code, std::string username);
+std::string generateVerificationEmailText(uint32_t puid, std::string code, std::string username);
 
 std::string encrypt(std::string str);
 std::string decrypt(std::string str);
